+++ examples/msw/src/gen/mocks/createCustomer.ts
@@ -10,7 +10,7 @@ import { faker } from '@faker-js/faker'
 export function createCustomer(data?: Partial<Customer>): Customer {
   faker.seed([220])
   return {
-    ...{ id: faker.number.int(), username: faker.string.alpha(), address: faker.helpers.multiple(() => createAddress()) as any },
+    ...{ id: faker.number.int(), username: faker.string.alpha(), address: faker.helpers.multiple(() => createAddress()) },
     ...(data || {}),
   }
 }