+++ packages/plugin-faker/src/generators/__snapshots__/enumNames.ts
@@ -5,5 +5,5 @@
 import { faker } from '@faker-js/faker'
 
 export function enumNamesType() {
-  return faker.helpers.arrayElement<any>(['Pending', 'Received'])
+  return faker.helpers.arrayElement<any>([0, 1])
 }